+++ src/components/generic/confirm-dialog.styles.ts
@@ -1,5 +1,5 @@
 import { createStyles } from "@material-ui/core";
-import { globalStyles } from "../../../styles/globalStyles";
+import { globalStyles } from "../../styles/globalStyles";
 
 export const styles = createStyles({
   ...globalStyles,
