+++ src/components/generic/error-dialog.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import strings from "../../../localization/strings";
+import strings from "../../localization/strings";
 import * as moment from "moment";
 import { Dialog, DialogTitle, DialogContent, DialogContentText, DialogActions, Button } from "@material-ui/core";
